Adjust for splitting refactor
