+++ packages/block-editor/src/components/block-list/use-block-props/use-selected-block-event-handlers.js
@@ -5,7 +5,7 @@ import { ENTER, BACKSPACE, DELETE } from '@wordpress/keycodes';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { useRefEffect } from '@wordpress/compose';
 import { createRoot } from '@wordpress/element';
-import { store as blocksStore, getDefaultBlockName } from '@wordpress/blocks';
+import { store as blocksStore } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -24,23 +24,14 @@ import BlockDraggableChip from '../../../components/block-draggable/draggable-ch
  */
 export function useEventHandlers( { clientId, isSelected } ) {
 	const { getBlockType } = useSelect( blocksStore );
-	const {
-		getBlockRootClientId,
-		isZoomOut,
-		hasMultiSelection,
-		getBlockName,
-		canInsertBlockType,
-		getNextBlockClientId,
-		getBlockOrder,
-		getBlockEditingMode,
-	} = unlock( useSelect( blockEditorStore ) );
+	const { getBlockRootClientId, isZoomOut, hasMultiSelection, getBlockName } =
+		unlock( useSelect( blockEditorStore ) );
 	const {
 		insertAfterBlock,
 		removeBlock,
 		resetZoomLevel,
 		startDraggingBlocks,
 		stopDraggingBlocks,
-		selectBlock,
 	} = unlock( useDispatch( blockEditorStore ) );
 
 	return useRefEffect(
@@ -77,49 +68,13 @@ export function useEventHandlers( { clientId, isSelected } ) {
 					return;
 				}
 
-				event.preventDefault();
-
-				if ( keyCode === ENTER && isZoomOut() ) {
-					resetZoomLevel();
-				} else if ( keyCode === ENTER ) {
-					const rootClientId = getBlockRootClientId( clientId );
-					if (
-						canInsertBlockType(
-							getDefaultBlockName(),
-							rootClientId
-						)
-					) {
-						insertAfterBlock( clientId );
-					} else {
-						function getNextClientId( id ) {
-							let nextClientId = null;
-
-							while (
-								typeof id === 'string' &&
-								! ( nextClientId = getNextBlockClientId( id ) )
-							) {
-								id = getBlockRootClientId( id );
-							}
-
-							return nextClientId;
-						}
-
-						let nextClientId =
-							getBlockOrder( clientId )[ 0 ] ??
-							getNextClientId( clientId );
-
-						while (
-							nextClientId &&
-							getBlockEditingMode( nextClientId ) === 'disabled'
-						) {
-							nextClientId = getNextClientId( nextClientId );
-						}
-
-						if ( nextClientId ) {
-							selectBlock( nextClientId, null );
-						}
+				if ( keyCode === ENTER ) {
+					if ( isZoomOut() ) {
+						event.preventDefault();
+						resetZoomLevel();
 					}
 				} else {
+					event.preventDefault();
 					removeBlock( clientId );
 				}
 			}
